Added centralized error handling.

+++ ms-business/src/main/java/org/wscale/mountains/service/MountainServiceImpl.java
@@ -6,6 +6,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import org.wscale.commons.exceptions.ErrorObject;
+import org.wscale.commons.exceptions.errors.NotFoundRuntimeException;
 import org.wscale.mountains.domain.MountainEntity;
 import org.wscale.mountains.repositories.MountainRepository;
 
@@ -24,7 +26,13 @@ public class MountainServiceImpl implements MountainService{
 
     @Override
     public MountainEntity getMountain(final Long id) {
-        return this.mountainRepository.findOne(id);
+        MountainEntity mountainEntity = this.mountainRepository.findOne(id);
+        if (mountainEntity == null) {
+            throw new NotFoundRuntimeException(
+                    new ErrorObject("The mountain you have requested with the id "+id+" was not found.",
+                            "Make sure you provide a existing mountain id here."));
+        }
+        return mountainEntity;
     }
 
     @Override
@@ -39,6 +47,12 @@ public class MountainServiceImpl implements MountainService{
 
     @Override
     public void deleteMountain(final Long id) {
+        MountainEntity loadedMountain = this.mountainRepository.findOne(id);
+        if (loadedMountain == null) {
+            throw new NotFoundRuntimeException(
+                    new ErrorObject("The mountain you want to delete with the id "+id+" does not exist.",
+                            "Make sure you provide a existing mountain id here."));
+        }
         this.mountainRepository.delete(id);
     }
